add server setup
