+++ server.js
@@ -1,10 +1,42 @@
 #!/usr/bin/env node
 
+
+/**
+ * Set up server for app
+ */
+
+var express = require('express');
+var path = require('path');
+var bodyParser = require('body-parser');
+
+var app = express();
+
+// view engine setup
+app.set('views', path.join(__dirname, 'view'));
+app.set('view engine', 'hbs');
+
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(express.static(path.join(__dirname, 'public')));
+
+var placeHolder = ""
+
+app.get('/', function(req, res) {
+ res.redirect('/flickrtwit')
+})
+
+app.get('/flickrtwit', function(req, res) {
+ res.render('touch', placeHolder)
+})
+
+app.get('/flickrtwit/:id', function(req,res){
+  
+})
+
 /**
  * Module dependencies.
  */
 
-var app = require('./index');
 var http = require('http');
 
 /**
@@ -86,3 +118,5 @@ function onListening() {
     ? 'pipe ' + addr
     : 'port ' + addr.port;
 }
+
+module.exports = app;
